Added a lot of classes to the SpectraCyber and restructured existing entities and controllers to handle abstract implementations.
