+++ ControlRoomApplication/ControlRoomApplication/Entities/SpectraCyber/SpectraCyberModeTypeEnum.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ControlRoomApplication.Entities
+{
+    public enum SpectraCyberModeTypeEnum
+    {
+        UNKNOWN,
+        CONTINUUM,
+        SPECTRAL
+    }
+}
