+++ ControlRoomApplication/ControlRoomApplication/Entities/SpectraCyber/SpectraCyberSimaultor.cs
@@ -1,0 +1,7 @@
+﻿namespace ControlRoomApplication.Entities
+{
+    public class SpectraCyberSimulator : AbstractSpectraCyber
+    {
+        public SpectraCyberSimulator() : base() { }
+    }
+}
